feat(backend): migrate to playwright + template changes
We build and bundle chromium via playwright for each platform now.

This is to make it less "weird" to download the backend and then
have puppeteer download other assets as well. Which is "scary"
from a security standpoint.

Also we want to be in full control of the releases and don't have to
rely on external download sources, other than the release assets
hoster itself.

Just imagine you download a new binary, but can't use it,
because whatever service that is serving the builds for playwright
or puppeteer is down.

It makes it also way nice from a UX perspective to just download
the backend with feedback of progress.

Previously it was just a "still downloading..." and
even worse "still installing" for ages, when puppeteer was installing
all the assets it needs.

The only downside to this approach is, that we need to
implement signing and notarizing for apple asap.

Otherwise, I'm pretty sure that at least the chromium binary won't work.